services/compat-dify/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sevenflows-compat-dify"
version = "0.1.0"
description = "7Flows compat:dify adapter stub service."
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi>=0.115.0,<1",
  "PyYAML>=6.0.2,<7",
  "pydantic-settings>=2.6.0,<3",
  "uvicorn[standard]>=0.32.0,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
